requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
rank-bm25>=0.2.2
numpy>=1.22
pypdf>=4.0
python-pptx>=1.0
reportlab>=4.0
python-dotenv>=1.0
python-multipart>=0.0.9
loguru>=0.7
pydantic>=2.0
pymupdf>=1.24
sqlalchemy>=2.0
beautifulsoup4>=4.12
squareup
